Split cleanup logic into a separate class
