+++ service-api/src/main/java/greencity/service/files/StorageCleanupService.java
@@ -1,0 +1,8 @@
+package greencity.service.files;
+
+public interface StorageCleanupService {
+    /**
+     * Method for scheduled storage cleanup.
+     */
+    void cleanUp();
+}
